[IMP] spreadsheet: allow collaborative public sharing
Some people still use other well known spreadsheet software instead of
o-spreadsheet because they want to share a spreadsheet with edit rights to
external people. It's currently not possible with odoo spreadsheet.

task-6121503
