+++ addons/spreadsheet/__manifest__.py
@@ -11,11 +11,11 @@
         'views/public_readonly_spreadsheet_templates.xml',
     ],
     'assets': {
-        'web.chartjs_lib' : [
+        'web.chartjs_lib': [
             'spreadsheet/static/lib/chartjs-chart-geo/chartjs-chart-geo.js',
             'spreadsheet/static/lib/chart_js_treemap.js',
         ],
-        'spreadsheet.o_spreadsheet': [
+        'spreadsheet.o_spreadsheet_core': [
             'web/static/src/views/graph/graph_model.js',
             'web/static/src/views/pivot/pivot_model.js',
             'web/static/src/polyfills/clipboard.js',
@@ -26,79 +26,36 @@
             'spreadsheet/static/src/o_spreadsheet/o_spreadsheet.xml',
             'spreadsheet/static/src/**/*.xml',
             ('remove', 'spreadsheet/static/src/assets_backend/**/*'),
-            ('remove', 'spreadsheet/static/src/public_readonly_app/**/*'),
+            ('remove', 'spreadsheet/static/src/public_spreadsheet/**/*'),
+        ],
+        'spreadsheet.o_spreadsheet': [
+            ('include', 'spreadsheet.o_spreadsheet_core'),
         ],
         'web.assets_web_print': [
             'spreadsheet/static/src/print_assets/**/*',
         ],
         'spreadsheet.public_spreadsheet': [
-            ('include', 'web.assets_frontend_minimal'),
-            ('include', 'web._assets_helpers'), # bootstrap variables
-            'web/static/src/scss/bootstrap_overridden.scss',
-            'web/static/src/scss/pre_variables.scss',
-            'web/static/lib/bootstrap/scss/_variables.scss',
-            'web/static/lib/bootstrap/scss/_variables-dark.scss',
-            'web/static/lib/bootstrap/scss/_maps.scss',
-            ('include', 'web._assets_bootstrap'),
-            'web/static/lib/popper/popper.js',
-            'web/static/lib/bootstrap/js/dist/util/index.js',
-            'web/static/lib/bootstrap/js/dist/dom/data.js',
-            'web/static/lib/bootstrap/js/dist/dom/event-handler.js',
-            'web/static/lib/bootstrap/js/dist/dom/manipulator.js',
-            'web/static/lib/bootstrap/js/dist/dom/selector-engine.js',
-            'web/static/lib/bootstrap/js/dist/util/config.js',
-            'web/static/lib/bootstrap/js/dist/base-component.js',
-            'web/static/lib/bootstrap/js/dist/collapse.js',
-            'web/static/lib/bootstrap/js/dist/dropdown.js',
-            ('include', 'web.icons_fonts'),
-            'web/static/lib/owl/owl.js',
-            'web/static/lib/luxon/luxon.js',
-            'web/static/src/libs/luxon.js',
-            'web/static/src/owl2/owl3_compatibility_layer.js',
-            'web/static/src/owl2/utils.js',
-            'web/static/lib/owl/odoo_module.js',
-            'web/static/src/core/utils/**/*.js',
-            'web/static/src/core/browser/browser.js',
-            'web/static/src/core/browser/feature_detection.js',
-            'web/static/src/core/registry.js',
-            'web/static/src/core/assets.js',
-            'web/static/src/core/templates.js',
-            'web/static/src/core/template_inheritance.js',
-            'web/static/src/session.js',
-            'web/static/src/env.js',
-            'web/static/src/core/**/*.js',
-            ('remove', 'web/static/src/core/emoji_picker/emoji_data.js'),
-            ('include', 'web.chartjs_lib'),
-            'spreadsheet/static/src/o_spreadsheet/o_spreadsheet.js',
-            'spreadsheet/static/src/o_spreadsheet/o_spreadsheet.xml',
+            ('include', 'web.assets_frontend'),
+            ('include', 'spreadsheet.o_spreadsheet_core'),
+            'web/static/src/model/**/*.js',
+            'web/static/src/views/**/*.js',
+            'web/static/src/search/**/*.js',
+            'web/static/src/core/commands/command_hook.js',
+            'web/static/src/webclient/menus/menu_helpers.js',
+            ('remove', 'spreadsheet/static/src/ir_ui_menu/spreadsheet_link_service.js'),
             'spreadsheet/static/src/o_spreadsheet/o_spreadsheet.css',
-            'spreadsheet/static/src/o_spreadsheet/icons.xml',
-            'spreadsheet/static/src/o_spreadsheet/o_spreadsheet_extended.scss',
-            'spreadsheet/static/src/o_spreadsheet/migration.js',
-            'spreadsheet/static/src/helpers/odoo_functions_helpers.js',
-            'spreadsheet/static/src/o_spreadsheet/odoo_module.js',
-            'spreadsheet/static/src/helpers/helpers.js',
-            'spreadsheet/static/src/o_spreadsheet/cancelled_reason.js',
-            'spreadsheet/static/src/logging/**/*',
-            'spreadsheet/static/src/public_readonly_app/**/*.xml',
-            'spreadsheet/static/src/public_readonly_app/**/*.scss',
-            'spreadsheet/static/src/public_readonly_app/**/*',
-            'spreadsheet/static/src/hooks.js',
-            'spreadsheet/static/src/plugins.js',
+            'spreadsheet/static/src/**/*.scss',
+            'spreadsheet/static/src/public_spreadsheet/**/*',
         ],
         'web.assets_backend': [
             'spreadsheet/static/src/o_spreadsheet/o_spreadsheet.css',
             'spreadsheet/static/src/**/*.scss',
             'spreadsheet/static/src/assets_backend/**/*',
-            ('remove', 'spreadsheet/static/src/public_readonly_app/**/*.scss'),
             ('remove', 'spreadsheet/static/src/print_assets/**/*'),
         ],
         'web.assets_unit_tests': [
             'spreadsheet/static/tests/**/*',
             ('include', 'spreadsheet.o_spreadsheet'),
-            'spreadsheet/static/src/public_readonly_app/**/*.xml',
-            'spreadsheet/static/src/public_readonly_app/**/*.js',
-            ('remove', 'spreadsheet/static/src/public_readonly_app/main.js'),
         ],
     }
 }
