+++ addons/spreadsheet/static/src/ir_ui_menu/ir_ui_menu_plugin.js
@@ -13,6 +13,10 @@ export class IrMenuPlugin extends OdooCorePlugin {
      * @returns {object | undefined}
      */
     getIrMenu(menuId) {
+        if (!this.env.services.menu) {
+            // menu service isn't available in the public bundle
+            return undefined;
+        }
         let menu = this.env.services.menu.getMenu(menuId);
         if (!menu) {
             menu = this.env.services.menu.getAll().find((menu) => menu.xmlid === menuId);
