--- addons/spreadsheet/static/src/public_readonly_app/public_readonly.scss
@@ -1,55 +1,0 @@
-$filter-panel-width: 300px;
-
-#spreadsheet-mount-anchor {
-  background-color: #DDDDDD;
-  overflow: hidden;
-  display: flex;
-  flex-direction: row-reverse;
-}
-.o-spreadsheet {
-  font-size: 12px;
-  height: 100%;
-  flex-grow: 1;
-
-  .o-topbar-toolbar .o-readonly-toolbar.text-muted {
-      color: $info !important;
-  }
-}
-
-.o-public-spreadsheet-filters {
-  background-color: #E7E9ED;
-  gap: 20px 70px;
-  width: $filter-panel-width;
-
-  .o-public-spreadsheet-filters-close-button {
-    width: 20px;
-    height: 20px;
-    cursor: pointer;
-  }
-}
-
-.o-public-spreadsheet-filter-button {
-  z-index: 100;
-  cursor: pointer;
-
-  &:hover {
-    background-color: #E7E9ED;
-  }
-}
-
-.o-public-spreadsheet:has(.o-spreadsheet-print) {
-  .o-public-spreadsheet-filter-button,
-  .o-public-spreadsheet-filters {
-    display: none !important;
-  }
-}
-
-
-@media print {
-  .o-public-spreadsheet {
-    height: 100% !important;
-  }
-  #spreadsheet-mount-anchor {
-    overflow: unset;
-  }
-}
