+++ addons/spreadsheet_dashboard/static/src/public/public_dashboard.scss
@@ -1,0 +1,22 @@
+$filter-panel-width: 300px;
+
+.o-public-spreadsheet-filters {
+  background-color: #e7e9ed;
+  gap: 20px 70px;
+  width: $filter-panel-width;
+
+  .o-public-spreadsheet-filters-close-button {
+    width: 20px;
+    height: 20px;
+    cursor: pointer;
+  }
+}
+
+.o-public-spreadsheet-filter-button {
+  z-index: 100;
+  cursor: pointer;
+
+  &:hover {
+    background-color: #E7E9ED;
+  }
+}
