+++ addons/spreadsheet_dashboard/static/tests/tours/share_dashboard_public_tour.js
@@ -1,0 +1,10 @@
+import { registry } from "@web/core/registry";
+
+registry.category("web_tour.tours").add("spreadsheet_share_dashboard_public", {
+    steps: () => [
+        {
+            trigger: ".o-spreadsheet",
+            content: "spreadsheet is loaded",
+        },
+    ],
+});
