+++ addons/spreadsheet_dashboard/tests/test_share_dashboard_tour.py
@@ -11,10 +11,8 @@ class TestDashboardShareTour(DashboardTestCommon, HttpCase):
         """check the public spreadsheet page can be opened without error"""
         dashboard = self.create_dashboard()
         share = self.share_dashboard(dashboard)
-        # web_tour is not part of the public dashboard assets bundle.
-        # We can't use the start_tour helper method.
-        self.browser_js(
-            "/dashboard/share/%s/%s" % (share.id, share.access_token),
-            "console.log('test successful');",
-            ready="odoo.isReady",
+        self.start_tour(
+            '/dashboard/share/%s/%s' % (share.id, share.access_token),
+            'spreadsheet_share_dashboard_public',
+            login=None,  # public user
         )
